+++ readme/RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 * [#23](https://github.com/skuzzle/snapshot-tests/issues/33): Allow to configure strictness of JSON comparison
-* [#24](https://github.com/skuzzle/snapshot-tests/issues/34): Support for HTML snapshots
+* [#34](https://github.com/skuzzle/snapshot-tests/issues/34): Support for HTML snapshots
+* [#36](https://github.com/skuzzle/snapshot-tests/issues/36): Throw proper `AssertionError` if actual snapshot input is null
 * Build against JUnit 5.9.1 (coming from 5.8.2)